Move TodoDirective to directives.rs
