+++ crates/ruff/src/checkers/mod.rs
@@ -4,5 +4,4 @@ pub(crate) mod imports;
 pub(crate) mod logical_lines;
 pub(crate) mod noqa;
 pub(crate) mod physical_lines;
-pub(crate) mod todo_directives;
 pub(crate) mod tokens;
