+++ ruff.schema.json
@@ -2368,6 +2368,13 @@
         "SLF00",
         "SLF001",
         "T",
+        "T",
+        "T0",
+        "T00",
+        "T001",
+        "T002",
+        "T003",
+        "T004",
         "T1",
         "T10",
         "T100",
